requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
fastapi==0.115.8
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
natsort==8.4.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.8.0
pytest==8.3.4
pytest-mock==3.14.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.27.1
pika==1.3.2
